middleware/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.32.3
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
ezkl==16.2.3
pydantic==2.10.4
email_validator==2.2.0